requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and Machine Learning
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
alembic>=1.11.0  # Database migrations

# API and HTTP
httpx>=0.24.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Trading APIs (choose based on broker)
# kiteconnect>=4.0.0  # Zerodha Kite Connect
# fyers-apiv2>=2.0.0  # Fyers API
# alice-blue>=2.0.0   # Alice Blue API

# Configuration and logging
pyyaml>=6.0
python-json-logger>=2.0.0
colorlog>=6.7.0

# Async and concurrency
asyncio>=3.4.3
aiofiles>=23.0.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Scheduling and automation
schedule>=1.2.0
apscheduler>=3.10.0

# Utilities
click>=8.1.0  # CLI interface
tqdm>=4.65.0  # Progress bars
python-dateutil>=2.8.0
pytz>=2023.3

# Security
cryptography>=41.0.0

# Optional: Telegram bot for notifications
# python-telegram-bot>=20.0.0

# Optional: Email notifications
# sendgrid>=6.10.0

# Development dependencies (optional)
jupyter>=1.0.0
ipython>=8.14.0 